+++ R/form_draft_detail.R
@@ -1,0 +1,61 @@
+#' Show details of a Draft Form.
+#'
+#' `r lifecycle::badge("experimental")`
+#'
+#' The response includes standard overall Form metadata, like
+#' `xmlFormId`, in addition to the Draft-specific information.
+#'
+#' @template param-pid
+#' @template param-fid
+#' @template param-url
+#' @template param-auth
+#' @template param-retries
+#' @return A list with the Draft Form's metadata as per the ODK Central
+#'   API docs, including the `draft_token`.
+#'   Top level list elements are renamed from ODK's `camelCase` to
+#'   `snake_case`.
+# nolint start
+#' @seealso \url{https://docs.getodk.org/central-api-form-management/#getting-draft-form-details}
+# nolint end
+#' @family form-management
+#' @export
+#' @examples
+#' \dontrun{
+#' # See vignette("setup") for setup and authentication options
+#' # ruODK::ru_setup(svc = "....svc", un = "me@email.com", pw = "...")
+#'
+#' fl <- form_list()
+#'
+#' d <- form_draft_detail(fid = fl$fid[[1]])
+#'
+#' d$xml_form_id
+#' }
+form_draft_detail <- function(
+  pid = get_default_pid(),
+  fid = get_default_fid(),
+  url = get_default_url(),
+  un = get_default_un(),
+  pw = get_default_pw(),
+  retries = get_retries()
+) {
+  yell_if_missing(url, un, pw, pid = pid, fid = fid)
+
+  httr::RETRY(
+    "GET",
+    httr::modify_url(
+      url,
+      path = glue::glue(
+        "v1/projects/{pid}/forms/",
+        "{URLencode(fid, reserved = TRUE)}/draft"
+      )
+    ),
+    httr::add_headers("Accept" = "application/json"),
+    httr::authenticate(un, pw),
+    times = retries
+  ) |>
+    yell_if_error(url, un, pw) |>
+    httr::content(encoding = "utf-8") |>
+    janitor::clean_names()
+}
+
+# usethis::use_test("form_draft_detail")  # nolint